requirements.txt
Flask
gunicorn
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
requests
google-auth-oauthlib==0.4.6
google-auth==2.3.0
google-api-python-client==2.22.0
urllib3==1.26.7
celery
redis
schedule

